Implemented DownloadService

+++ src/org/digitalcampus/oppia/service/DownloadService.java
@@ -7,16 +7,23 @@ import android.preference.PreferenceManager;
 import android.support.v4.content.LocalBroadcastManager;
 import android.util.Log;
 
+import com.splunk.mint.Mint;
+
 import org.digitalcampus.mobile.learning.R;
 import org.digitalcampus.oppia.activity.PrefsActivity;
 import org.digitalcampus.oppia.utils.storage.FileUtils;
 
 import java.io.File;
+import java.io.FileOutputStream;
 import java.io.IOException;
+import java.io.InputStream;
 import java.net.HttpURLConnection;
 import java.net.MalformedURLException;
 import java.net.ProtocolException;
 import java.net.URL;
+import java.security.DigestInputStream;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
 
 public class DownloadService extends IntentService {
@@ -77,13 +84,14 @@ public class DownloadService extends IntentService {
 
         String fileUrl = intent.getStringExtra(SERVICE_URL);
         String filename = intent.getStringExtra(SERVICE_FILENAME);
-        String digest = intent.getStringExtra(SERVICE_DIGEST);
+        String fileDigest = intent.getStringExtra(SERVICE_DIGEST);
+        File downloadedFile = null;
 
         try {
             URL url = new URL(fileUrl);
             //If no filename was passed, we set the filename based on the URL
             if (filename == null){ filename = url.getPath().substring(url.getPath().lastIndexOf("/")+1); }
-            File file = new File(FileUtils.getMediaPath(this), filename);
+            downloadedFile = new File(FileUtils.getMediaPath(this), filename);
 
             HttpURLConnection connection = (HttpURLConnection) url.openConnection();
             connection.setRequestMethod("GET");
@@ -105,28 +113,72 @@ public class DownloadService extends IntentService {
                 return;
             }
 
-            //Download the file
-            while (true) {
+            FileOutputStream f = new FileOutputStream(downloadedFile);
+            InputStream in = connection.getInputStream();
+
+            MessageDigest mDigest = MessageDigest.getInstance("MD5");
+            in = new DigestInputStream(in, mDigest);
+
+            byte[] buffer = new byte[8192];
+            int len1;
+            long total = 0;
+            int previousProgress = 0, progress = 0;
+            while ((len1 = in.read(buffer)) > 0) {
                 //If received a cancel action while downloading, stop it
                 if (isCancelled(fileUrl)) {
-                    deleteFile(file);
+                    deleteFile(downloadedFile);
                     removeCancelled(fileUrl);
                     break;
+                }
+
+                total += len1;
+                progress = (int)((total*100)/fileLength);
+                if ( (progress > 0) && (progress > previousProgress)){
+                    sendBroadcast(fileUrl, ACTION_DOWNLOAD, ""+progress);
+                }
+                f.write(buffer, 0, len1);
+            }
+            f.close();
+
+            if (fileDigest != null){
+                // check the file digest matches, otherwise delete the file
+                // (it's either been a corrupted download or it's the wrong file)
+                byte[] digest = mDigest.digest();
+                String resultMD5 = "";
+
+                for (int i=0; i < digest.length; i++) {
+                    resultMD5 += Integer.toString( ( digest[i] & 0xff ) + 0x100, 16).substring( 1 );
+                }
+                if(!resultMD5.contains(fileDigest)){
+                    this.deleteFile(downloadedFile);
+                    sendBroadcast(fileUrl, ACTION_FAILED, this.getString(R.string.error_media_download));
+                    return;
                 }
             }
 
         } catch (MalformedURLException e) {
             e.printStackTrace();
+            sendBroadcast(fileUrl, ACTION_FAILED, this.getString(R.string.error_media_download));
+            return;
         } catch (ProtocolException e) {
+            this.deleteFile(downloadedFile);
             e.printStackTrace();
+            sendBroadcast(fileUrl, ACTION_FAILED, this.getString(R.string.error_media_download));
+            return;
         } catch (IOException e) {
+            this.deleteFile(downloadedFile);
             e.printStackTrace();
+            sendBroadcast(fileUrl, ACTION_FAILED, this.getString(R.string.error_media_download));
+            return;
+        } catch (NoSuchAlgorithmException e) {
+            Mint.logException(e);
+            e.printStackTrace();
+            sendBroadcast(fileUrl, ACTION_FAILED, this.getString(R.string.error_media_download));
+            return;
         }
 
         sendBroadcast(fileUrl, ACTION_COMPLETE, null);
-
     }
-
 
     /*
     * Sends a new Broadcast with the results of the action
@@ -162,7 +214,7 @@ public class DownloadService extends IntentService {
     }
 
     private void deleteFile(File file){
-        if (file.exists() && !file.isDirectory()){
+        if ((file != null) && file.exists() && !file.isDirectory()){
             file.delete();
         }
     }
